logbook/backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   starlette
    #   watchfiles
click==8.1.7
    # via
    #   -r requirements.in
    #   uvicorn
fastapi==0.109.2
    # via
    #   -r requirements.in
    #   strawberry-graphql
graphql-core==3.2.3
    # via strawberry-graphql
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
httptools==0.6.1
    # via uvicorn
idna==3.6
    # via anyio
psycopg==3.1.18
    # via -r requirements.in
pydantic==2.6.2
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.16.3
    # via pydantic
pydantic-settings==2.2.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via strawberry-graphql
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.9
    # via
    #   -r requirements.in
    #   strawberry-graphql
pyyaml==6.0.1
    # via uvicorn
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via anyio
sqlalchemy==2.0.29
    # via -r requirements.in
starlette==0.36.3
    # via fastapi
strawberry-graphql[fastapi]==0.224.1
    # via -r requirements.in
typing-extensions==4.9.0
    # via
    #   fastapi
    #   psycopg
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   strawberry-graphql
uvicorn[standard]==0.27.1
    # via -r requirements.in
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
